+++ src/OpenTelemetry/Trace/TracerProviderExtensions.cs
@@ -49,8 +49,8 @@ namespace OpenTelemetry.Trace
         /// </summary>
         /// <param name="provider">TracerProviderSdk instance on which ForceFlush will be called.</param>
         /// <param name="timeoutMilliseconds">
-        /// The number of milliseconds to wait, or <c>Timeout.Infinite</c> to
-        /// wait indefinitely.
+        /// The number (non-negative) of milliseconds to wait, or
+        /// <c>Timeout.Infinite</c> to wait indefinitely.
         /// </param>
         /// <returns>
         /// Returns <c>true</c> when force flush succeeded; otherwise, <c>false</c>.
@@ -68,13 +68,13 @@ namespace OpenTelemetry.Trace
                 throw new ArgumentNullException(nameof(provider));
             }
 
+            if (timeoutMilliseconds < 0 && timeoutMilliseconds != Timeout.Infinite)
+            {
+                throw new ArgumentOutOfRangeException(nameof(timeoutMilliseconds), timeoutMilliseconds, "timeoutMilliseconds should be non-negative or Timeout.Infinite.");
+            }
+
             if (provider is TracerProviderSdk tracerProviderSdk)
             {
-                if (timeoutMilliseconds < 0 && timeoutMilliseconds != Timeout.Infinite)
-                {
-                    throw new ArgumentOutOfRangeException(nameof(timeoutMilliseconds), timeoutMilliseconds, "timeoutMilliseconds should be non-negative.");
-                }
-
                 try
                 {
                     return tracerProviderSdk.OnForceFlush(timeoutMilliseconds);
@@ -95,8 +95,8 @@ namespace OpenTelemetry.Trace
         /// </summary>
         /// <param name="provider">TracerProviderSdk instance on which Shutdown will be called.</param>
         /// <param name="timeoutMilliseconds">
-        /// The number of milliseconds to wait, or <c>Timeout.Infinite</c> to
-        /// wait indefinitely.
+        /// The number (non-negative) of milliseconds to wait, or
+        /// <c>Timeout.Infinite</c> to wait indefinitely.
         /// </param>
         /// <returns>
         /// Returns <c>true</c> when shutdown succeeded; otherwise, <c>false</c>.
@@ -115,13 +115,13 @@ namespace OpenTelemetry.Trace
                 throw new ArgumentNullException(nameof(provider));
             }
 
+            if (timeoutMilliseconds < 0 && timeoutMilliseconds != Timeout.Infinite)
+            {
+                throw new ArgumentOutOfRangeException(nameof(timeoutMilliseconds), timeoutMilliseconds, "timeoutMilliseconds should be non-negative or Timeout.Infinite.");
+            }
+
             if (provider is TracerProviderSdk tracerProviderSdk)
             {
-                if (timeoutMilliseconds < 0 && timeoutMilliseconds != Timeout.Infinite)
-                {
-                    throw new ArgumentOutOfRangeException(nameof(timeoutMilliseconds), timeoutMilliseconds, "timeoutMilliseconds should be non-negative.");
-                }
-
                 if (Interlocked.Increment(ref tracerProviderSdk.ShutdownCount) > 1)
                 {
                     return false; // shutdown already called